backend/requirements.txt
fastapi
uvicorn
PyJWT
sqlalchemy[asyncio]
asyncpg
bcrypt
python-dotenv
anthropic
python-multipart
google-genai
httpx
supabase
pypdf
python-docx
pgvector
